requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
enum-tools==0.13.0
frozenlist==1.7.0
idna==3.10
multidict==6.6.3
propcache==0.3.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.4
six==1.17.0
twitchAPI==4.5.0
typing_extensions==4.14.1
urllib3==2.5.0
vk-api==11.9.9
yarl==1.20.1
